Fix UT

+++ modules/Nncase.Modules.CPU/Evaluator/CPU/Im2col.cs
@@ -43,13 +43,13 @@ public sealed class Im2colEvaluator : ITypeInferencer<Im2col>, ICostEvaluator<Im
         int pad_h_after = target.Padding[1];
         int pad_w_before = target.Padding[2];
         int pad_w_after = target.Padding[3];
-        int kernel_h = target.Kernel[0];
-        int kernel_w = target.Kernel[1];
+        long kernel_h = target.Kernel[0];
+        long kernel_w = target.Kernel[1];
         int stride_h = target.Stride[0];
         int stride_w = target.Stride[1];
-        int output_h = ((height + pad_h_before + pad_h_after -
+        long output_h = ((height + pad_h_before + pad_h_after -
                 ((1 * (kernel_h - 1)) + 1)) / stride_h) + 1;
-        int output_w = ((width + pad_w_before + pad_w_after -
+        long output_w = ((width + pad_w_before + pad_w_after -
          ((1 * (kernel_w - 1)) + 1)) / stride_w) + 1;
         var outputTensor = new float[inChannel * kernel_h * kernel_w * batch * output_h * output_w * lanes];
 
@@ -149,13 +149,13 @@ public sealed class Im2colEvaluator : ITypeInferencer<Im2col>, ICostEvaluator<Im
         int pad_h_after = target.Padding[1];
         int pad_w_before = target.Padding[2];
         int pad_w_after = target.Padding[3];
-        int kernel_h = target.Kernel[0];
-        int kernel_w = target.Kernel[1];
+        long kernel_h = target.Kernel[0];
+        long kernel_w = target.Kernel[1];
         int stride_h = target.Stride[0];
         int stride_w = target.Stride[1];
-        int output_h = ((height + pad_h_before + pad_h_after -
+        long output_h = ((height + pad_h_before + pad_h_after -
                 ((1 * (kernel_h - 1)) + 1)) / stride_h) + 1;
-        int output_w = ((width + pad_w_before + pad_w_after -
+        long output_w = ((width + pad_w_before + pad_w_after -
          ((1 * (kernel_w - 1)) + 1)) / stride_w) + 1;
         return tt with { Shape = new Dimension[] { tt.Shape[1] * kernel_h * kernel_w, tt.Shape[0] * output_h * output_w } };
     }
