+++ modules/Nncase.Modules.CPU/IR/CPU/Functional.cs
@@ -87,12 +87,12 @@ public partial class CPU
         return new Call(new ResizeImage(packedAxes, padedNums, newSize, resizeMode, transformationMode, nearestMode), input);
     }
 
-    public static Expr Im2col(Expr input, int[] kernel, int[] stride, int[] padding)
+    public static Expr Im2col(Expr input, long[] kernel, int[] stride, int[] padding)
     {
         return new Call(new Im2col(kernel, stride, padding, Array.Empty<int>(), Array.Empty<int>()), input);
     }
 
-    public static Expr Im2col(Expr input, int[] kernel, int[] stride, int[] padding, int[] packedAxes, int[] padedNums)
+    public static Expr Im2col(Expr input, long[] kernel, int[] stride, int[] padding, int[] packedAxes, int[] padedNums)
     {
         return new Call(new Im2col(kernel, stride, padding, packedAxes, padedNums), input);
     }
